requirements.txt
matrix-nio
openai-whisper
pydub
requests
torch
tdqm
aiofiles